GuestInfo: Let users tweak poll loop behavior.
Requested nearly a year ago by Ken Barr, this change introduces a new
Tools config option for the GuestInfo plugin.  It is

- guestinfo.poll-interval:  an integer which defines the
GuestInfoGather poll interval in seconds.  A value of 0 disables
the poll loop.

This is the first plugin making use of the new tcs_conf_reload signal.
Changes to the Tools config file emit this signal.  In response to
receiving it, iff the poll interval has changed, the GuestInfoGather
timeout source is canceled, and a new one with the new interval is
created and attached.

A small side effect of this change is the addition of a macro,
CONFGROUPNAME_GUESTINFO, which specifies the config file group for
GuestInfo variables.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>