Original Description: Add RPC to support enabling and disabling 3D in guest.
Add an RPC that allows the UI to reset Aero -> essentially net
stop/start uxsms. Programatically done by DCE Win32 APIs. See
http://msdn.microsoft.com/en-us/library/aa969538%28VS.85%29.aspx.
This is basically a fix for 467709 until a better solution becomes
available (one that requires methods that are not supported
right now, like PCI hotplug).

The fix is a bit ugly, we need to refactor a bit in the code that
spawns resolution set.

Furthermore, we need to broadcast this to all user sessions, not
just the active one. A loop that iterates all guest sessions will
come later.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>