Run programs in a user environment without Tools' settings.
vmware-user on Linux uses the app-loader tool to set up its
environment so it can use various bundled libraries. However,
this cause these environment variables to be set that when
VIX runs programs or scripts from vmware-user, these processes
fail due to library conflicts. For example, gedit on RHEL 4 was
compiled with a version of libpango that conflicts with the
version of this library that ships with the Tools.

The solution is to store the original, pre-app-loader
environment with the VIX tools module. Some of the
infrastructure was already in place to fix a similar issue when
launching a web browser to view the Help from toolbox. However,
this needed to be plumbed through VIX, and added to lib/procMgr.

Also, one further wrinkle is that VIX allows clients to
programatically change the environment. So, we keep track of the
environment using a hash table, update it when processing
WriteVariable commands, then serial the hashtable into a environ
block when RunProgram or RunScript are called.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>