Add log-only miRegionStr() utility to lib/region
I know lib/region is much-maligned and will go away, but I needed some
debugging besides the Warning()s provided by miPrintRegion(), so I
made this change.

With this change, I added a new VMX86_LOG-only function, miRegionStr(),
which allocates a string describing the contents of a region, and
stores the string inside the region structure.

When the region is changed or destroyed, the string is freed.

I consolidated the region initalization code duplicated in
miRegionCreate and miRegionInit into miRegionInit (I accidentally
modified only miRegionInit initially, and realized it was silly
to have two functions that do the exact same thing, so I figured
I'd save future coders the same trouble.)

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>