Add unique MODULE_ALIAS to non-PCI linux drivers
To help installer figure out whether there is already a vendor-provided
kernel module (with potentially different name) that should not be
clobbered it was decided that we will use unique module alias string to
identify modules.

vmsync driver will not be upstreamed (newer kernels already have the
same functionality) and so is vmblock and thus are left alone.
vmxnet, vmxnet3, pvscsi and vmci are PCI drivers and don't require
artificial aliases.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>