Remove rpchChannel.h's dependendy on rpcin.h.
rpcChannel.h is part of the "public" tools core API, and shouldn't expose
internal APIs (such as those in rpcin.h). This has the downside of having
to duplicate some definitions from rpcin.h, since we still have code that
uses the rpcin library directly.

This change also standardizes rpcChannel.h's interface to use glib types
and not VMware types. This is the cause of most of the churn in this
change (Bool -> gboolean), and also required the definition of a new type
(RpcInRet) to be used while we still have code that needs to be compatible
both with vmtoolsd and vmware-user; this type can (and should) go away
once vmware-user is deprecated.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>