Re-organize the Tools Core public headers (1/2).
These headers will form the base of the public Tools Core interface, so
they need to be moved to the "fully qualified" location.

This change just moves the headers to the new location. (For the
perforce-minded folks, I am branching / editing the new files, so
history is kept as far as where the contents come from.)

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>