XdrUtil: Add optional string, opaque field macros.
During the NicInfoV3 mess, it was a fairly common operation to have to
assign arbitrary data to optional string, opaque, and optional opaque
fields.  This change introduces a few macros which make it a little easier
to do just that.

Not sure that I'm happy with them, but they get the job done for initializers
used in test code.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>