Vmblock-FUSE: respond to reads from control file
Because we can't implement the same control mechanism in FUSE-based
vmblock module as in-kernle one (FreeBSD and Solaris use ioctl and
we don't have a device node here) userspace needs to be able to
detect the kind of vmblock being used and adjust accordingly. Change
vmblock-fuse to allow reading from the control file and respond with
a pre-defined string ("I am VMBLOCK-FUSE").

Ths change complements change 816902 that implemented fuse/legacy
selector in DND code.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>