GuestInfo: Fix NicInfo caching. [1/2]
NicInfoV3 introduced considerably more complex XDR types.  Comparing
said types is non-trivial, so this change introduces "isEqual" routines
for each of the new types.

The IsEqual code bodies are similar to some of the existing test code,
and the latter will be refactored to use these soon.

I also have a bunch of unit tests coming in a second change.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>