This change was reviewed as multiple pieces but is being checked in as one big chunk to prevent build breaks since parts depend on other pieces.
Add VMCILOG to the guest

The guest and the host crosstalk drivers have different notions of logging macros. Really we should unify them.
To keep this change from getting any larger I have instead added a temporary VMCILOG macro to the guest.
====
Add VMCI_CanScheduleDelayedWork and VMCI_ScheduleDelayedWork to all VMCI platforms

This change is a precursor to supporting VMCI delayed event callbacks. It provides basic
infrastructure to ask that a callback be scheduled and run in a delayed work context.
Right now this is only implemented on linux and vmkernel. Other platforms currently provide
dummy implementations and return VMCI_ScheduleDelayedWork() as FALSE.
====
Add support for delayed linux VMCI events.

This is the intial change to add delayed event callbacks to vmciEvent.c.
In particular this change:
1. Defines a flags field and requires that all callers pass in delayed or none flags when calling
VMCIEvent_RegisterSubscription.
2. Changes the lock rank of the event callbacks to be above queuepairs.
3. Change the VMCIEvent cleanup routine to require all VMCI events to be unregsitered before the module is unloaded.
4. Reference count the individual vmci event subscriptions so they don't go away while callbacks can be firing.
5. Wait for all callbacks to fire before allowing unsubscription to continue.
6. Modify VMCIEvent_Dispatch to allow callbacks to be fired in a delayed work context with no locks held.
====
Pull in vmciEvent.c changes from the host into vmciEvent.c

This change is a setup change to unify vmciEvent.c on the guest and the host module. This change just pulls in
host side vmciEvent.c code into the guest. Followup changes will make this code actually work.
====
Remove HelperQueue VMKERNEL call from vmciEvent.c.

Previous to this change the VMkernel went through the vmci event subscriber array in a helper world because it
wanted to take the subscriber array lock which was lower rank than the queuepair lock. Now that the VMK always uses
the delayed work mechanism ... this is no longer needed.
====
Make vmkernel work with the new vmciEvent code.

This modifies the core vmkernel vmci interface code to use the new vmci event flag field. This lets
vmkernel continue to compile and lets external (to the module) vmkernel clients use vmci if they desire.
====
Merge bora-vmsoft vmciEvent.c into bora vmciEvent.c

Now that the vmciEvent code is in a state where it can work on both the guest and the host, merge the guest
vmciEvent.c code to be 100% the same as the host code. In the future we can consider just using the same file
for the host and guest.
====
Add extra VMCIEvent_Subscribe param to vsock linux build.

This change just modifies the linux vsock code to pass in the new required param for Event subscribe to existing
calls. This should not change anything about how linux vsock actually works.
====
Register for VMCI_EVENT_CTX_ID_UPDATE

This change makes linux vsock register for VMCI_EVENT_CTX_ID_UPDATE and use the new vmciEvent delayed callback code.
Right now stream sockets break if a vm's cid ever changes and vsock is not reloaded (new stream sockets will not
succeed in being connected). This change fixes that by registering for the VMCI_EVENT_CTX_ID_UPDATE call and
reregistering the datagram handler for the control port.
====
Make common vsock code pass in the new argument to VMCIEvent_Subscribe

This change is required to have the common vsock code continue building. There should be no functional changes
here.
====

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>