Configure core dump limit on POSIX automatically.
In this change:

. if core dumps are enabled in the config file, try to raise the core
rlimit for the process when logging is enabled; this can still fail
if the system is configured to now allow it, but it's better than
not getting core dumps because the default system setting is to set
the limit to 0.

. on POSIX systems, try to chdir to the user's home if trying to create
a core dump, but the current directory is not writable.

. don't enable the Win32 exception filter automatically when loading the
vmtools DLL; instead, do it when configuring logging and requesting
core dumps.

. avoid recursive log calls from CoreDump_CoreDump() on Win32; only Panic()
was incrementing the recursive panic count, it needs to be done in the
actual core dump function.

This change doesn't handle SIGSEGV on POSIX; that will still create a core
dump if the limit is set appropriately, but won't go through the code that
sees if the current directory is writable.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>