Factor out reading guest and host time from TimeSyncDoSync.
This change splits out reading guest and host time from the main
time sync loop.  At the same time, it converts to using "diff" and
"interruptLag" to "gosError" and "apparentError".  The polarity of
these two now match NTP/adjtime: positive values mean that time is
ahead, negative values mean that time is behind.  The names match
TimeTracker.

This change shouldn't introduce any behavioral changes.  The
change in how errors are represented makes some of the conditions
in the one time sync portion of the loop look a bit suspect, but
I'll address that later.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>