char is neither signed nor unsigned
As some of our code starts to be compiled on ARM, we need to improve
some elements of code cleanliness.  This review deals with char being
neither signed nor unsigned in the gcc type system.  (Confused?  See
this link:)
http://www.trilithium.com/johan/2005/01/char-types/
This was initially found with gcc, but MSVC also makes the distinction.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>