Add unrestricted local encoding support to tools
Tools needs to be able to use local encodings even if they are
not supported in our stripped down version of the ICU.  In tools
builds accept any encoding as valid provided we recognize it
and it is the default local encoding.  If the encoding is not
supported in our ICU, revert to using native OS conversions with
codesetOld.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>