Allow plugins to respond to registration errors.
This change adds a callback that plugins can register to be notified of
registration errors. Plugins can use this functionality to selectively
disable functionality based on the current state of the container where
they are running.

For example, if a plugin depends on a signal provided by another plugin,
but this other plugin fails to load for any reason, the first plugin is
notified that the signal connection failed, and can disable the
functionality that relied on that signal.

Plugins are never unloaded, though. So they should still be careful to not
crash even if registration errors happen.

Bonus: clarify a few spots in the API docs.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>