FreeBSD VMBLOCK: avoid crash when doing lookup in a subdirectory
When trying to reconstruct path to a vnode during lookup of a file
within subdirectory the driver gets confused and crashes the kernel.
This can be easily triggered by doing H->G DnD of a small directory.
Instead of trying to reconstruct it every time let's just store the
latest lookup that lead to discovery of the vnode (similar to how
Linux driver does this).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>