Add support for datagram handles with VMCI_FLAG_ANYCID_DG_HND.
In certain situations (e.g. vsock) ... we really want to register datagram handles that are not
tied to one specific context id. The goal is that these datagram handles will continue to be
able to send and recv datagrams after a cid change. Without this change clients would have to go
through and manually reregister all datagram handles.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>