Timesync: fix backward sync when config file is reloaded
Reloading of tools configuration file causes purging of all options,
including ones that are sent via RPC by VMX. This causes issues with
timesync plugin, which 'forgets' that it is supposed to perform full
time synchronization when enabling time synchronization loop.

Solve the problem by storing value of TOOLSOPTION_SYNCTIME_STARTUP
in a private variabe instead of retrieving it from the config
dictionary.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>