vmtoolsd: do not store VMX options with the rest of config
VMX options are not persistent in the sense that they are not stored
in tools config file in GOS but rather [re]transmitted from host at
startup, resume, etc, or when they are changed. When we reload config
file (which we periodically do to switch to the latest and greatest
user-supplied settings) VMX options get flushed from memory so let's
not put them in key file in first place and instead require plugins
to store values relevant to their operation themselves.

The only user who did look up VMX option in key file - timeSync - has
already been changed to store value by itself.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>