Remove bogus error checks from af_vsock.c
With the new VMADDR_CID_ANY change, a datagram handle that is registered with VMCI will have a context
of invalid on the guest or hosted context id on the host. This means the error check in linux vsock
is no longer correct and was preventing sockets from working on the host. Just remove the checks since
it is really a bug if VMCI ever returns us a handle with bad values and VMCI_SUCCESS...

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>