Hostinfo:  Add OS name caching to POSIX impl.
While investigating PR 479137, we noticed that the
Hostinfo_GetOSName may end up holding up the Tools service for a while
on a heavily loaded host.  My theory is that we're starved of either
CPU or disk, and this change brings the caching available in the Windows
implementation over to the POSIX side.

After the first invocation, we won't have to go to disk or fork any
other processes in order to identify the running operating system.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>