Allow hgfsServer to use glib mutexes instead of bora/lib/lock.
This avoids having lib/lock and its dependencies having to be part of
open-vm-tools. To make sure vmware-user still works in open-vm-tools,
make sure the glib thread library is initialized, even though it's not used
by the non-open-vm-tools version.

This change only really affects the open-vm-tools build at the moment;
eventually we might want to clean up the regular tools build also, so
that we have one less dependency on bora.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>