Rework the vmrpcdbg library to use CUnit.
CUnit provides nicer ways to track test failures than just logging or
using hard asserts. So use CUnit to run vmtoolsd-based test code.

This required a few changes:

. to be able to use CUnit, the vmtoolsd main loop needs to be run in
the context of a CUnit test. So the code in mainLoop.c was reorganized
a bit so that all code that touches plugin code is run in that context,
and execution of the main loop is done through the vmrpcdbg library
(which sets up the CUnit context) when running in debug mode.

. the interface of the vmrpcdbg library was cleaned up a little bit.

. test code was changed to use CUnit asserts instead of
ASSERT/g_assert/logs.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>