Try to auto-detect CUnit without having to define CUSTOM flags.
Like in other cases where there's no pkg-config or config scripts,
use a default to try to find the library. Add a new macro that does
that in a generic way and use it in other cases where that situation
occurs in configure.ac.

Bonus: remove libnotify detection code since none of the open-vm-tools
code uses libnotify.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>