TimeSync: cleanup hanlding of the timesync loop state
We used to store current state of plugin/timeloop in timeSyncState;
unfortunately we were also using presence of 'timer' to infer whether
the loop is running which is confusing. Make 'state' the pimary source
of this information.

Also we should not stop timesync loop any time sync fails, let's treat
treat this as a transient condition and keep running.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>