Switch to using a more generic loaded-module-list func
There is a not-very-generic loaded-module-list function
in lib/user (which is Linux-only).  For other reasons,
I've written a more generic one in lib/coreDump
(checked in with a previous review, cn970593) that works
for Linux, Windows, and Mac.

Delete the old one, and hook up the one callsite to call the
new one.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>