update to I18N encoding name cross reference table
Update the I18N name to encoding cross reference table with
some new names.  Identify encodings that are supported by ICU
even if we don't support them in our stripped down ICU data
file.  This doesn't actually add any encodings, but it makes
it easier should we decide to do so in the future.

Add new encodings that are needed for [361930], and change the
ICU data path to point to the new location in toolchain.

Also add some test code that can be enabled from Unicode_Init to
verify the entries in the xRef table match what is expected in
the ICU data file.  Add a similar change to codeset to disable
getting the ICU data file from toolchain if a custom file is
being tested.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>