Use Solaris version number to decide MAC_VERSION to be used.
The earlier fix used devo_rev to choose the MAC_VERSION to be used in
mac_alloc(). That was a workaround for incomplete header files list in
toolchain. Moreover it did not serve the purpose as devo_rev came from
our header files itself and not as a runtime parameter from Solaris
kernel.

The header files for Opensolaris 2009.06 in toolchain have been restored	. This change exploits the same. Some macros and structures have
changed in Opensolaris since Solaris 10. Depending on the version for
which the driver is being compiled, the right ones are used.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>