Add Str_Strerror, a thread safe version of strerror
strerror is not thread safe, it reuses the memory for the error
string on every call.  strerror_r is thread safe, but requires
callers to manage their buffers.  Add a new Str_Strerror function
that uses thread local storage to keep the error string.  This
function is only present on GLIBC_VERSION_23 and later.

The change to vmm-link-vmm-to-static.make undefines GLIB_VERSION_23,
which was erroneously set.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>