Add guest kernel-mode support to vmci on Windows
* Added new IOCTLs to create/destroy datagram handles from
kernel-mode (the existing ones rely on a process context).

* Removed the old shared memory IOCTLs and code.

* Modified the guest driver to handle the new IOCTLs.

* Modified the host driver to fail gracefully when called from
kernel-mode (since not yet supported).

* Added kernel-mode vmcilib that wraps the IOCTLs.  Interface
is exactly the same as if you linked against vmci directly.

* Added a kernel-mode version of crosstalktest for all of the
above.  Note that it is built out of bora but will run in the
guest too; it will pass on the guest but fail in the host until
support is added.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>