Exclude pagemap from in-guest vm-support
Newer kernels, such as those included in Ubuntu 9.04
and 9.10, export a 'pagemap' node to the proc filesystem.
This node maps guest virtual to guest physical pages,
allowing address look-ups for the entire valid guest
virtual address space (including unmapped regions).
Thus, archiving the entire pagemap for each process/task
is prohibitively large (not to mention extremely slow) on
64-bit systems.  More importantly, this information is
not terribly useful for our debugging.  So, we exclude it
and vm-support in Ubuntu 9.x guests won't automatically
fill the guest disk.  (See also PR 399587 for the
linux host vm-support fix for the same issue).

As the 'Collect support data...' button now runs this
script in-guest, this change is important.  (Previously,
to trigger this problem, customers would have had to run
vm-support manually inside the guest.)

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>