Revised version of cs 979736
cs 979736 included changes to lib/string to call GetReturnAddress()
that caused the linker to build differently.  cs 979736 was
subsequently backed out for pr 483076.  This is the same change
without the GetReturnAddress() calls.  This change should not
introduce any functional differences whatsoever; it merely introduces
a new library function in lib/string that is currently not called.

The original description of cs 979736 is:

Add Str_Strerror, a thread safe version of strerror

strerror is not thread safe, it reuses the memory for the error
string on every call.  strerror_r is thread safe, but requires
callers to manage their buffers.  Add a new Str_Strerror function
that uses thread local storage to keep the error string.  This
function is only present on GLIBC_VERSION_23 and later.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>