vmtoolsd: fix console handling on Win32.
The previous code required the user to instruct the code to attach to the
console manually, which was error-prone. It also had a bug where the help
wouldn't show up, since the code to attach to the console was called after
the help was shown.

This change does three things:

. auto-detects the help triggers in the command-line and attaches a console
as needed.

. make sure we have a console when command-line parsing fails, so that the
message is shown.

. attaches a console when configuring logging, whenever a log domain is
configured to output to the terminal.

As a result, the old "--console" command line option is not needed anymore.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>