Allow outbound-only RpcChannels, cleanup API.
. allow a channel to be set up for outbound operations only. This defers
creation of an RpcIn channel to until its needed, and doesn't enforce
that data for RpcIn operations is available when starting the channel.

This allows clients who only want to send outbound messages to more easily
use the channel. Before, because of the code enforcing the call to the
setup() function, these apps could run into issues like failing to properly
start the channel because of some conflict caused by the RpcIn channel name
(among other possible issues).

. make the private data in the RpcChannel structure private to the
lib/rpcChannel code.

This only exposes the bare minimum of the structure in the header file to
allow for new implementations of RpcChannel.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>