Add timeout to backup operation.
This change makes the Tools backup components mimic the VC behavior and
cancel the operation after 15 min (configurable in the case of Tools).

Previously, VC would time out the operation but Tools would never be
notified, so they'd happily keep waiting for VSS to finish its thing.
This would cause problems for customers, who would then retry to take
the quiesced snapshot just to get a "I'm still busy" message from Tools.

To abort the operation, the old support for canceling backup operations
in the vmbackup plugin had to be fixed a little bit (it was pretty much
a no-op in the previous state, so was very untested). It should now be
doing the right thing, and the code is simplified a little bit as a result.

The VSS components themselves don't handle aborting the operation optimally;
they just check from time to time (i.e., at the end of the "logical blocks"
of the backup operation) whether an abort was requested, and stop if it is.
So there may still be a small window between VC timing out and Tools timing
out, but we'll have to live with that.

Bonus: fix a regression that caused app quiescing to almost always be
disabled on Windows 2003.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>