Toolbox: fix black icon from Linux guests in Unity mode
When switching to Unity mode icon for toolbox in Linux guests
is mangled thusly: transparent pixels left as transparent and
filled pixels are turned into black ones. With old logo it was
not noticeable since boxes were done with transparent pixels so
that logo outline was preserved. With the new logo consisting
only of colored pixels (with the exception of couple transparent
ones in the corners) the result is solid black square in place
of the icon, which is not acceptable.

The logo is currently attached to the window with gdk_window_set_icon.
GDK documentation does not recommend using this function, suggesting
gdk_window_set_icon_list() instead, which, coincidentally, fixes
the color mangling issue.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>