Unity/X11: Avoid infinite loops.
When fielding a moveresize request, the guest is expected to return the
window's resulting coordinates, even if the request failed.  The Unity/
X11 code was written such that after issuing the moveresize request to
the X server, it'd recursively & manually pump the event loop until
receiving a configurenotify event for the window in question.
Unfortunately, there is no guarantee that such an event will ever
arrive.

Since our implementation is synchronous with respect to the X server,
that logic isn't at all needed.  Instead, we'll simply flush the X
channel and make sure all requests have been handled (using XSync),
and query the X server again for the new coordinates.

Now, in theory, this could introduce some latency to the moveresize
operation, but correctness trumps performance right now.  If
performance is ever important, this code will need a rewrite, anyway.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>