Reorganize the logging code in vmtoolslib.
This change is a preparation for some enhancements I'll work on for the
file loggers. It does the following:

. separate each logger implementation into their own source file.
. provide an abstraction for loggers so that, if wanted, we can add new
logger implementations in the future (e.g., syslog) more easily.

The main goal, though, is just to better organize the code and separate
the logger implementations from the common logging code.

This change does introduce a change from the previous behavior: log files
are always written in UTF-8. stdout and OutputDebugString logs are, like
before, in the local user encoding.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>