Add a message catalog infrastructure to vmtoolslib.
It is based on the Msg_* infrastructure but modified to make it easier to
use in Tools. The main differences are:

. it only provides message catalog-related functionality, no message queues
or anything like that.
. it supports more languages (Msg_* seems to be restricted to en/ja on
Win32), and has fallbacks (e.g. en_US -> en) similar to how gettext
does it.
. properly implements message domains so that different binaries (i.e.,
plugins) can be loaded into the same process and have different message
catalogs, avoiding message id conflicts.
. the message catalog format extends the dictionary format to allow for
continuation lines, so that long messages can be broken into multiple
lines in the catalog file.
. it doesn't use Dictionary_* functions which are way too encumbered.
. it provides easier accessors for wide strings (Win32 only) (at the cost
of some memory overhead).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>