Make the tools file logger thread-safe.
This change uses a rw lock to protect the shared data in the logger. This
allows a multi-threaded environment to safely use file loggers, even with
the current features it has (lazy open of log files, log file rolling).

Note this does not protect multiple threads from writing to the file at
the same time (we just "leak" the thread-safety characteristics of the
underlying system call, if any).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>