User level multi-threading fixes for gethostbyname
gethostbyname is not always thread safe.  While it uses
thread local storage on some platforms, such as Mac, it
does not on Linux.  Add a non-Windows version of
Posix_GetHostByName and use it where appropriate.  Also
replace some code calling uname and gethostbyname in
HostInfoPopulateHostname with a call to Hostinfo_HostName,
since all it was trying to determine was the host name.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>