Add VMCIQPair interface to VMCI
VMCIQPair is a new interface that hides the queue pair internals.
Rather than access each queue in a pair directly, operations are now
performed on the queue as a whole.  This is simpler and less error-prone,
and allows for future queue pair features to be added under the hood
with no change to the client code.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>