Fix sync driver support in vmbackup plugin.
The cause of the hangs seen while investigating the bug is that the Legato
sync driver doesn't like even when people are doing read-only operations
like stat(2); so it would block the calling thread, which just happened to
be the Tools service's main thread, and things would go to hell.

The fix is to add a flag to the shared app context that lets code knows
when it's safe to access the disk. The two instances of the hang I was
able to capture in the kernel debugger (the log file check task and the
guest info gathering task) now check this flag and just bail out if it's
set. This is the "quick & less intrusive" version of the fix, which will
be modified once it hits vmcore-main to be more generic and more in line
with the Tools Core API.

This *may not* fix all instances of such hangs. I've actually been able
to hang a VM while performing a quiesced snapshot with this fix, but the
frequency is much, much, much lower (it happened once in a few hundred
tries). I haven't yet been able to catch this hang in a debugger. But,
code-wise, we're at least on par with K/L (in fact, better, since K/L
does not have the guest info workaround).

A flag was added that allows disabling the sync driver without having to
uninstall it, in case of emergency. The driver is enabled by default.

Also fix the sync driver support in the vmbackup plugin. The propagation
of the sync driver handle through the different operations was busted, so
querying the handle status was not working as expected. Also make sure
resources are properly freed and errors are correctly handled.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>