Rework the "disksFrozen" hack to be more generic.
CS 1010838 added a hack to work around an issue with the Win32 sync driver.
The hack is simpler than the more generic solution, which is fine for
the release branch it was targeted at, but changing the public plugin
API for such things is in general a bad idea.

This change implements it so that the current (i.e. before that patch)
public API is maintained but the same functionality can be achieved. It
requires a little bit more code, though, since it requires creating a
new gobject signal to notify of the state change.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>