vmware-user: Eliminate racy exit situation.
Vmware-user's signal handler was destroying RPC channels when it
shouldn't have.  Rather than having a few cases where we decide to
call VMwareUserCleanupRpc before exiting the main loop, this change
simply places a call to VMwareUserCleanupRpc /after/ the main loop
has exited.  (The X I/O error handler is a special case, as apps
are -not- supposed to return from that handler.)

In the case of bug 542135, vmware-user received SIGUSR2 while still
initializing, destroying gRpcIn, before control was returned to the
initialization and eventmanager routines which expected gRpcIn to
still exist.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>