Unity/X11: Work area struts are [x1,x2) and [y1,y2).
The region code defines a 25x25 rectangle with its upper left corner
at the origin as (x1,x2,y1,y2) = (0,25,0,25); upper bounds are
exclusive.  However, the strut code works with inclusive coordinates,
so we should instead use (0,24,0,24).  Without doing this, we run
into problems.

Consider a multihead setup of 2 displays at 1600x1200 in a horizontal
configuration.  There should be a single 25px top strut attached to the
left display.  By adding a pixel, we have a strut that ranges from
[0,1600], occupying a single pixel on the right display.  The guest will
now reserve the top 25px on the right display while the host does not.

Should someone maximize a window on that display, the guest places
the window's upper left corner at (0,25) while the host places it at
(0,0), and all hell breaks loose.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>