Linux Guest driver support for VMCI doorbell notifications
This change introduces three new functions to the VMCI guest
kernel API dealing with the HW v8 support for doorbell
notifications. The actual implementation of these doorbell
calls are embedded in a new notification module, that handles
allocating notification bitmap indexes,
registration/unregistration of callbacks for a given index,
and firing the handlers for a given index.

The Linux device initialization now opts in to use
notification, if the hardware supports it.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>