Unity/X11: Ignore decoration widget windows.
KDE 4's kwin provides decoration windows by creating independent windows
of _NET_WM_WINDOW_TYPE_NORMAL and reparenting them to a top-level/frame
window.  This window and the application's client window become siblings.

The problem is that Unity/X11 never had this case in mind.  Whenever we
saw a new top-level window, we expected it to either (a) stay as a top-
level window / become an override-redirect window, or (b) be reparented
to a different window and receive the WM_STATE property (which identifies
*this* window as the client window).

Usually a window manager would create its decorations immediately as
children of the frame window, not graft the decorations into the frame's
tree.  This caused Unity/X11 to treat these windows independently,
expecting them to eventually become a client window.  This might not
normally be a problem, but as window coordinates are relative to one's
parent, Unity/X11 thought these windows' coordinates were relative to the
root window rather than some regular frame window.

A correct fix for this would be to associate this window and its frame,
such that once the frame is ever associated with a client window, any
remaining client window "candidates" would be ignored.  (They are, after
all, unimportant children of the frame.  The only child relevant to us
is the application's client window.)  Unfortunately I don't have enough
time to work on that, so I'm instead sticking in a safe hack in order
to meet C&C RC stability requirements.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>