lib/file: tolerate Windows better
Windows doesn't have a way of being deterministically correct
when handling files as some file operations may return a "busy"
even when not no thread/process is operating on the targetted
file. This is due to the parent directory being used by another
thread/process at the time of the file operation.

The only recourse is to retry. The good news is these problems
are rare. Unfortunately we're seeing more of this on Vista++ than
in the past.

The old code did a fixed number of retries while it slightly
wobbled the amount of time slept between retries to try to
fight its way through these kinds of problems. This proved
insufficient.

The new way is to fully wobble the retries, using a highly variable
sleep duration after a failed attempt and bounding the retries based
on the total time waited rather than a fixed retry count.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>