suidWrapper: Ignore SIGUSR1, SIGUSR2.
Ubuntu 9.10+'s Upstart may result in the services script and
vmware-user launching concurrently.  When this happens, the services
script may deliver SIGUSR1 and/or SIGUSR2 to vmware-user while it's
still executing inside the suidWrapper or appLoader.  Neither were
handling SIGUSR1/2, and so they'd terminate.

This change sets the handler to SIG_IGN for these signals.
Note that SIG_IGN persists across fork() and execve(), so we
don't need to block the handles again inside the appLoader driver
(toolbox/linux/vmwareuser/loader).

As a setuid application, classically we avoided pulling in other
libraries, but the vmsignal library is safe.  It's tiny and doesn't
use bora/bora-vmsoft at all, so its use here is okay.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>